@@ -7,14 +7,15 @@
     "babel-loader": "^7.1.1",
     "babel-plugin-transform-react-constant-elements": "^6.23.0",
     "babel-plugin-transform-react-inline-elements": "^6.22.0",
-    "babel-preset-env": "1.6.0",
+    "babel-preset-env": "1.6.1",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-2": "^6.24.1",
-    "happypack": "^4.0.0-beta.2",
-    "lint-staged": "^4.0.3",
+    "happypack": "^4.0.1",
+    "lint-staged": "^7.0.0",
     "pre-commit": "^1.2.2",
-    "prettier": "^1.5.3",
+    "prettier": "^1.11.1",
     "webpack": "^3.5.5",
+    "webpack-bundle-analyzer": "^2.10.0",
     "webpack-manifest-plugin": "^1.3.1"
   },
   "scripts": {
@@ -22,11 +23,12 @@
     "build:sitemaps": "./scripts/sitemaps.js",
     "build:production": "npm run build && npm run build:sitemaps",
     "start": "webpack --watch",
-    "lint:staged": "lint-staged"
+    "lint:staged": "lint-staged",
+    "format": "yarn prettier --write '**/*.{js,md,scss}'"
   },
   "lint-staged": {
-    "*.js": [
-      "prettier --write --single-quote --trailing-comma=es5",
+    "*.{js,md,scss}": [
+      "prettier --write",
       "git add"
     ]
   },
@@ -38,21 +40,24 @@
   "private": true,
   "dependencies": {
     "@haroenv/react-sparklines": "^1.7.1",
-    "algoliasearch": "^3.24.3",
+    "algoliasearch": "^3.24.9",
     "bootstrap": "4.0.0-alpha.5",
-    "bytes": "^2.5.0",
-    "date-fns": "^2.0.0-alpha.1",
-    "docsearch.js": "^2.4.1",
-    "jquery": "^3.2.1",
-    "marked": "^0.3.6",
-    "prop-types": "^15.5.10",
+    "bytes": "^3.0.0",
+    "date-fns": "^2.0.0-alpha.7",
+    "docsearch.js": "^2.5.2",
+    "jquery": "^3.3.1",
+    "marked": "^0.3.17",
+    "prop-types": "^15.6.1",
     "qs": "^6.5.0",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
-    "react-instantsearch": "^4.0.10",
-    "react-transition-group": "1.2.0",
+    "react": "^16.1.1",
+    "react-dom": "^16.1.1",
+    "react-instantsearch": "^5.0.0",
+    "react-transition-group": "^1.2.1",
     "unescape-html": "^1.0.0",
     "unfetch": "^3.0.0",
     "xss": "^0.3.3"
+  },
+  "resolutions": {
+    "algoliasearch": "3.25.1"
   }
 }
